Created stack.py
Your task is to extend the Stack class behavior in such a way so that the class is able to count all the elements that are pushed and popped (we assume that counting pops is enough). Use the Stack class we've provided in the editor.

Follow the hints:

introduce a property designed to count pop operations and name it in a way which guarantees it is hidden;
initialize it to zero inside the constructor;
provide a method which returns the value currently assigned to the counter (name it get_counter()).